Update patch addressing further comments
