+++ core/src/main/scala/org/apache/spark/SparkContext.scala
@@ -415,7 +415,7 @@ class SparkContext(config: SparkConf) extends Logging with ExecutorAllocationCli
 
     if (master == "yarn" && deployMode == "client") System.setProperty("SPARK_YARN_MODE", "true")
     if (_conf.get(SPARK_IO_ENCRYPTION_ENABLED) && !SparkHadoopUtil.get.isYarnMode()) {
-      throw new SparkException("IO encryption is only supported in Yarn mode, please disable it " +
+      throw new SparkException("IO encryption is only supported in YARN mode, please disable it " +
         "by setting spark.io.encryption.enabled to false")
     }
